% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBAMSeq1.R
\name{NBAMSeq1}
\alias{NBAMSeq1}
\title{Differential Expression Analysis Based on Negative Binomial Additive Model}
\usage{
NBAMSeq1(
  object,
  gamma = 2.5,
  BPPARAM = NULL,
  use_bam = FALSE,
  AIC = FALSE,
  alpha_bound = Inf,
  save_model = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{A \code{\link[=NBAMSeqDataSet]{NBAMSeqDataSet()}} object.}

\item{gamma}{A number greater than or equal to 1. Increase gamma to create
smoother models. Default gamma is 2.5. See \code{\link[mgcv:gam]{mgcv::gam()}} for details.}

\item{BPPARAM}{An argument provided to \code{\link[BiocParallel:bplapply]{BiocParallel::bplapply()}}. See
\code{\link[BiocParallel:register]{BiocParallel::register()}} for details.}

\item{use_bam}{Logical, whether to use \code{\link[mgcv:bam]{mgcv::bam()}} instead of \code{\link[mgcv:gam]{mgcv::gam()}}. \code{bam} can be faster for large data.}

\item{AIC}{Logical, whether to generate AIC/BIC for each gene.}

\item{alpha_bound}{Numeric, set upper bound for gene-wise dispersion parameter. Can help if there are errors. A value of 10 is sufficiently large.}

\item{save_model}{Logical, whether to save the final \code{\link[mgcv:gam]{mgcv::gam()}} model for each gene. Warning: will take up a lot of memory.}

\item{verbose}{Logical, whether to print verbose output.}

\item{...}{Additional arguments provided to \code{\link[mgcv:gam]{mgcv::gam()}} or \code{\link[mgcv:bam]{mgcv::bam()}}.}
}
\value{
A \code{\link[=NBAMSeqDataSet]{NBAMSeqDataSet()}} object with \code{metadata(object)$fitted == TRUE}.
}
\description{
This function performs differential expression analysis based
on a negative binomial additive model.
}
\examples{
gsd <- NBAMSeq::makeExample(n = 3, m = 10)
gsd <- NBAMSeq::NBAMSeq1(gsd)
}
\references{
Love, M.I., Huber, W., Anders, S. (2014) Moderated estimation of
fold change and dispersion for RNA-seq data with DESeq2. Genome Biology,
15:550. \url{https://doi.org/10.1186/s13059-014-0550-8}
}
