% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{NBAMSeqDataSet-class}
\alias{NBAMSeqDataSet-class}
\title{NBAMSeqDataSet class}
\description{
\code{NBAMSeqDataSet} is a class inherited from
\code{\link{SummarizedExperiment}}.
It is used to store the count matrix, colData, and design formula
in differential expression analysis.
}
\section{Slots}{

\describe{
\item{\code{design}}{a mgcv-type design formula}
}}

