% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTable1.R
\name{topTable1}
\alias{topTable1}
\title{Return Gene Wise Fit Statistics}
\usage{
topTable1(
  object,
  coef = NULL,
  number = 10,
  genelist = NULL,
  adjust.method = "BH",
  sort.by = "P.Value",
  p.value = NULL
)
}
\arguments{
\item{object}{An \code{\link[=NBAMSeqDataSet]{NBAMSeqDataSet()}} object that has been fitted with \code{\link[=NBAMSeq1]{NBAMSeq1()}}.}

\item{coef}{Name of the term. If NULL or mismatched, then a list of all terms will be returned.}

\item{number}{Number of rows to be returned after sorting. Defaults to 10.}

\item{genelist}{Return just the genes in this list. Expects a character vector.}

\item{adjust.method}{P.adj methods as detailed in \link[stats:p.adjust]{stats::p.adjust.methods}. Defaults to "BH".}

\item{sort.by}{Column to sort by. Defaults to "P.Value".}

\item{p.value}{Cut off for p-value; only rows with p-values smaller than this value are shown. Defaults to NULL.}
}
\value{
A data.frame with the statistics.
}
\description{
Convenient wrapper to interact with the fitted \code{\link[=NBAMSeqDataSet]{NBAMSeqDataSet()}} object to
extract the fit statistics.
}
