% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NBAMSeq1.R
\name{NBAMSeq1}
\alias{NBAMSeq1}
\title{Differential expression analysis based on negative binomial additive
model}
\usage{
NBAMSeq1(
  object,
  gamma = 2.5,
  BPPARAM = NULL,
  use_bam = FALSE,
  AIC = FALSE,
  ...
)
}
\arguments{
\item{object}{a NBAMSeqDataSet object}

\item{gamma}{a number greater or equal to 1. Increase gamma to create
smoother models. Default gamma is 2.5. See \code{\link[mgcv]{gam}} for
details.}

\item{BPPARAM}{an argument provided to \code{\link{bplapply}}. See
\code{\link[BiocParallel]{register}} for details.}

\item{use_bam}{use bam from mgcv instead of gam. bam can be faster for large data.}

\item{AIC}{generate AIC/BIC for each gene or not.}

\item{...}{additional arguments provided to \code{\link[mgcv]{gam}} or \code{\link[mgcv]{bam}}}
}
\value{
a NBAMSeqDataSet object
}
\description{
This function performs differential expression analysis based
on negative binomial additive model.
}
\examples{
gsd <- makeExample(n = 3, m = 10)
gsd <- NBAMSeq1(gsd)
}
\references{
Love, M.I., Huber, W., Anders, S. (2014) Moderated estimation of
fold change and dispersion for RNA-seq data with DESeq2. Genome Biology,
15:550. \url{https://doi.org/10.1186/s13059-014-0550-8}
}
